@@ -31,32 +31,31 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^7.2.15",
-    "@angular/cdk": "^7.3.7",
-    "@angular/common": "^7.2.15",
-    "@angular/compiler": "^7.2.15",
-    "@angular/core": "^7.2.15",
-    "@angular/forms": "^7.2.15",
-    "@angular/http": "^7.2.15",
-    "@angular/material": "^7.3.7",
-    "@angular/platform-browser": "^7.2.15",
-    "@angular/platform-browser-dynamic": "^7.2.15",
-    "@angular/router": "^7.2.15",
-    "@auth0/angular-jwt": "^2.1.0",
-    "@ngrx/effects": "^7.4.0",
-    "@ngrx/store": "^7.4.0",
-    "@ngrx/store-devtools": "^7.4.0",
-    "@swimlane/ngx-datatable": "^14.0.0",
-    "ajv": "^6.10.0",
-    "angular-password-strength-meter": "0.0.4",
-    "core-js": "^2.5.7",
-    "css-element-queries": "^1.2.0",
-    "cytoscape": "^3.5.7",
+    "@angular/animations": "^8.2.14",
+    "@angular/cdk": "^8.2.3",
+    "@angular/common": "^8.2.14",
+    "@angular/compiler": "^8.2.14",
+    "@angular/core": "^8.2.14",
+    "@angular/forms": "^8.2.14",
+    "@angular/material": "^8.2.3",
+    "@angular/platform-browser": "^8.2.14",
+    "@angular/platform-browser-dynamic": "^8.2.14",
+    "@angular/router": "^8.2.14",
+    "@auth0/angular-jwt": "^4.0.0",
+    "@ngrx/effects": "^8.6.0",
+    "@ngrx/store": "^8.6.0",
+    "@ngrx/store-devtools": "^8.6.0",
+    "@swimlane/ngx-datatable": "^16.0.2",
+    "ajv": "^6.11.0",
+    "angular-password-strength-meter": "^1.0.2",
+    "core-js": "^3.6.4",
+    "css-element-queries": "^1.2.3",
+    "cytoscape": "^3.14.0",
     "cytoscape-cola": "^2.3.0",
     "cytoscape-dagre": "^2.2.2",
     "cytoscape-spread": "^3.0.0",
-    "d3": "^5.9.2",
-    "elkjs": "^0.4.1",
+    "d3": "^5.15.0",
+    "elkjs": "^0.6.1",
     "hammerjs": "^2.0.8",
     "html2canvas": "^1.0.0-rc.5",
     "immutable": "^3.8.2",
@@ -64,42 +63,43 @@
     "lodash": "^4.17.15",
     "material-design-icons": "^3.0.1",
     "moment": "^2.24.0",
-    "mxgraph": "^3.9.12",
+    "mxgraph": "^4.1.0",
     "normalize.css": "^8.0.1",
     "ol": "^5.3.3",
-    "rxjs": "^6.5.2",
-    "tslib": "^1.9.3",
-    "zone.js": "^0.8.29",
+    "rxjs": "^6.5.4",
+    "tslib": "^1.10.0",
+    "zone.js": "^0.10.2",
     "zxcvbn": "^4.4.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.13.0",
-    "@angular/cli": "^7.3.9",
-    "@angular/compiler-cli": "^7.2.15",
+    "@angular-devkit/build-angular": "~0.803.25",
+    "@angular/cli": "^8.3.25",
+    "@angular/compiler-cli": "^8.2.14",
     "@types/d3": "^5.7.2",
     "@types/hammerjs": "^2.0.36",
-    "@types/jest": "^24.0.15",
-    "@types/node": "^10.5.6",
-    "@types/ol": "^5.3.0",
+    "@types/jest": "^25.1.2",
+    "@types/node": "^13.7.1",
+    "@types/ol": "^5.3.6",
     "@types/zxcvbn": "^4.4.0",
-    "codelyzer": "^5.0.1",
+    "codelyzer": "^5.2.1",
     "cypress": "^3.8.3",
-    "jest": "^24.8.0",
-    "jest-preset-angular": "^7.1.1",
+    "jest": "^25.1.0",
+    "jest-preset-angular": "^8.0.0",
     "jest-zone-patch": "0.0.10",
-    "ngentest": "^0.4.1",
-    "npm-audit-ci-wrapper": "^2.2.1",
+    "ngentest": "^1.4.0",
+    "node-sass": "^4.13.1",
     "protractor": "^5.4.2",
-    "rimraf": "^2.6.3",
+    "rimraf": "^3.0.2",
     "rxjs-tslint": "^0.1.7",
-    "rxjs-tslint-rules": "^4.23.1",
-    "stylelint": "^10.0.1",
-    "stylelint-config-standard": "^18.3.0",
-    "ts-jest": "^24.0.2",
-    "ts-node": "^7.0.0",
-    "tslint": "^5.16.0",
-    "tslint-config-standard": "^8.0.1",
-    "typescript": "3.2.4"
+    "rxjs-tslint-rules": "^4.28.3",
+    "stylelint": "^13.1.0",
+    "stylelint-config-standard": "^20.0.0",
+    "ts-jest": "^25.2.0",
+    "ts-node": "^8.6.2",
+    "tslint": "^6.0.0",
+    "tslint-config-standard": "^9.0.0",
+    "tsutils": "^3.17.1",
+    "typescript": "3.5.3"
   },
   "fclConfig": {
     "lastChange": "2019-06-12 11:25:17 +0200"
